スラロームゲイン調整中
新歓期前に動く状態を保存しておく．
ログデータの構造体化，運動系の関数が見直されている．走行距離のずれも改善．

+++ sensor.h
@@ -42,7 +42,7 @@
 		volatile uint16_t ad_res[5],ad_pre_res[5];
 		volatile uint16_t base_l, base_r;								//基準値を格納
 		volatile int16_t dif_l, dif_r;									//AD値と基準との差
-		volatile uint16_t volt_bat;									//電源電圧監視
+		volatile float volt_bat;									//電源電圧監視
 		volatile float cont_r,cont_l;										//壁制御の比例定数
 		volatile float Cont_kp[5];									//壁制御の場合分け
 		volatile float omega_G, angle_G,omega_G_rad,pre_omega_G;
@@ -62,7 +62,7 @@
 		extern volatile uint16_t ad_res[5],ad_pre_res[5];
 		extern volatile uint16_t base_l, base_r;
 		extern volatile int16_t dif_l, dif_r;
-		extern volatile uint16_t volt_bat;
+		extern volatile float volt_bat;
 		extern volatile float omega_G, angle_G,omega_G_rad,pre_omega_G;
 		
 		extern volatile float cont_r,cont_l;
